fix test import errors
