+++ chromatic/tests/__init__.py
@@ -15,3 +15,4 @@ from .test_summaries import *
 from .test_withmodel import *
 from .test_ok import *
 from .test_history import *
+from .test_spectra import *
